+++ dukes-javaee/src/main/java/eu/agilejava/dukes/RootResource.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package eu.agilejava.snoop;
+package eu.agilejava.dukes;
 
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
